Add Placement, move more stuff from double to float.
The usage of float is very slightly less convenient because you need to type 'f' after a numeric literal or cast the result of some Math functions that don't have an alternative in here. However, it interoperates much more easily with both libGDX and JDKGDXDS, which implements float-based collections but not double-based ones.