Finished up the insane DeepMap bug fixing.
There's also a bug I re-found in SerpentDeepMapGenerator that I think had been reported a while ago in SquidLib, but I couldn't reproduce at the time. By contorting the random values, I can get it to happen still in SquidLib, but it shouldn't happen in SquidSquad now. The fix should be backported shortly.