Figure out a weird infinite-loop bug.
DungeonProcessor checks that the dungeon it generated is a very small area, and if it is too small, it regenerates the dungeon. Before, it would reuse the same base dungeon and environment, which would cause DungeonProcessor to check the same tiny dungeon hundreds of times before running out of stack and crashing. Now, if it needs to regenerate, it makes a new dungeon each time with DungeonBoneGen, which almost certainly should make a working dungeon before getting close to another stack overflow.