Minor cleanup in DijkstraMap, TrigTools, ZoneOfInfluence.
I made ZoneOfInfluence no longer marked as Serializable, and I intend to remove as much Java serialization as I can from the library. It's a problem in many Java systems and it may be removed from Java altogether in a future release, so we should follow what many best-practices guides recommend and avoid it.