Got DawnlikeDemo mostly working.
Pathfinding still seems very wrong, and obstacles seem to be ignored. I hadn't been able to test DijkstraMap before, so maybe this will find a bug in that. It could also be related to not flipping the y axis, which I thought we no longer needed. Everything else seems good.