Remove java.io.Serializable everywhere possible.
It stays in the tests of squidold, because those need to be pretty much unchanged to test for compatibility. I think I cleared out everything else. The built-in Java serialization is a huge problem in "serious business" apps, but it could also be a legitimate issue for networked games that are susceptible to hacking.