OK, it was a jdkgdxds problem.
IntLongOrderedMap.get() and related methods slowed way down because of the pattern of keys -- most keys had hundreds of copies with identical low-16-bits, which, because jdkgdxds no longer mixes, is a huge problem for collisions. Switching to a custom ObjectLongOrderedMap with CoordObjectMap's place() method works at 60FPS with a new limited-area draw() method.