Update jdkgdxds to use its fix for equality in int/long-keyed maps.
This was found when comparing SpatialMaps for equality, and two equivalent maps were not treated as equal for two reasons. The zeroValue wasn't copied, which was one part, but also that same value was compared with != in equals(), even for Object values. This should all be fixed now, and the test in question passes.