New, simpler GumNoise has more variety.
It's also much closer to how Taffy is already implemented, and should be about the same speed (sway() was changed to sinTurns(), and a seed change was moved, but not much else).

This also makes PhantomNoise more-randomly rotate its basis vectors using the given seed for its Hasher. This affects Taffy and Gum as well.