Remove PEAR and FLAN from Noise.
Pear was very close to Simplex, but slower and with noticeable seams all over the noise. Flan repeated quite obviously when at a high frequency, forming a clear grid. The higher-dimensional highest-quality niche is still filled by Foam and Mutant, and Taffy has a nice curvy-bubbly look in high dimensions as well.