Something is still subtly wrong.
I'm not sure exactly where the problem is or even what it is exactly, but the randomly rotated poles just don't look the same as other points on the sphere (generated using other methods, ones that work). I changed matrixMultiply() because it might have been wrong before, but again I'm not sure.