This should make Phantom correct between dims.
Before, there was probably an issue due to how Hasher works normally -- since it usually hashes a full array, having extra "empty" items would change the result, and higher dimensions would have such extra items. Thankfully, Hasher already has a hash() method that allows limiting how many items it considers, and we can use that now to make sure getting 3D noise from a PhantomNoise with dim 3 is the same as one with dim 5. (If the seed is the same).