Left a large note about why the different noise...
Behavior is the way it is. It has to do with 3D inputs being sent to NoiseWrapper, which handles lacunarity here and then inside this block hands off to an INoise, and the INoise introduces dimensions w and u. To contrast, the other way has Noise get its w and u dimensions added at the start, so they get affected by the FBM or Ridged lacunarity changes and become "fizzier-looking."

I'm also trying CyclicNoise here -- it needs its frequency increased a fair amount to be useful in world map noise.