Experiment with some relief-map-like effects.
This also finds a noticeable discontinuity in 6D Simplex noise... It might be possible to fix it, who knows... https://i.imgur.com/la12ZIm.png shows the issue. Both sides use Noise with SIMPLEX_FRACTAL, and the right modifies the output with a one-liner function. In lower dimensions the right side looks continuous, but not in 6D (the left side isn't either in 6D, it just is harder to spot).