Get the remaining Interpolator s compared.
This renames bounceOut2 and the other numbers to bounce2Out, to match other names like pow2Out. This also applies to the "In" Interpolator s.

The cos() in libGDX seems to have quite a bit more error than it needs to... In digital, we don't do the half-unit offset, and this seems to reduce the error a lot, especially on anything aligned to multiples of 360/(2^14) degrees, such as 45 degrees.