Update TextraTypist to latest commit, fix stuff.
After 0.8.1, a commit removed the line in Font.enableShader() that reset the Batch color to white. This didn't have a visible change in TextraTypist's tests, but did fix someone's alpha-related logic and does seem to be more correct. The problem is, GlyphGrid relied on that reset, and some colors looked very strange without it being reset frequently. Things seem fine now.