Changes to Translator.
Here, we weaken SemiRandom even more, and make it so if it produces numerically-close results, it is more likely to produce the same value when selecting from a Language component like openingVowels or midConsonants. It does this just by copying the Language and sorting its various arrays, making it so when SemiRandom returns 4, 5, or 6, those are more likely to refer to the same item in some arrays (because there could be 3 or more of that item in a row).