I like this breaking change! Dijkstra Deques.
Instead of using an ObjectList that was frequently edited from its first index, rather than its last, DijkstraMap's path is now an ObjectDeque, allowing fast access to any index, and fast editing for both the start and end. The main downside is that it isn't a List, so you can't get a subList of it, but that was a bad idea before anyway. It allocated a lot.

Oh yeah, and we're on Gradle 8.2 now! It seems to work fine. I don't notice much that has changed, though maybe it's a bit faster to start up.