Remove some Java 8 APIs that RoboVM fails with.
I can't be totally sure if RoboVM currently does fail when it encounters java.util.function types, but jdkgdxds already removed all usage of these types, and it seems to work there now. There's drop-in replacements in funderby already. This also removes the not-especially-useful FunctionalInterface annotation for the same reason (RoboVM).