Prepare to move Interpolations to digital.
This meant changing a few things -- the probably-unused-everywhere methods getTagList() and getInterpolatorList() are gone, as are nullability annotations in that file. The private REGISTRY is a LinkedHashMap now instead of an ObjectObjectOrderedMap. Interpolator.apply() now does its bounds checking a little differently, though most Interpolators will produce the same results.