OK, interesting findings.
The float inputs that produce the largest results from gradCoord3D are very close to 1.0, but that means each of those vertices will be lerped with a float input that is very close to 0. In time, each vertex will be lerped with more inputs that are close to 0. Weirdly, when a float input is close to 1, the first parameter is essentially unused, or has minimal effect. That means the vertices close to 0 get used almost completely when an input like xf is close to 1...