Pondering my orb.
I'm not sure if it makes sense to generate unit vectors on-the-fly, but we might be able to do it using any of probit(), Ziggurat, or potentially logit(), which was added here. While probit() would be better distributed, logit() might be faster (three arithmetic ops on doubles and a call to Math.log()) than probit() (4 branches, two using Math.sqrt() and Math.log(), and all but the rarest using many multiplications).