Prototype uneven-mix syntax for DescriptiveColor.
It still needs documentation, but it should work with "brown^5 red^2 orange white", where any number (n) after a named color will be treated as if that name was repeated (n) times, for the purposes of weighting it against other colors in the mix. That is, the above color String is the same as "brown brown brown brown brown red red orange white". All other punctuation works in place of ^ (except for the underscore, which can be used in color names).