Get rid of the weird ((char)1) usage.
There wasn't very much of it left! This appeared as either `'\1'` or the number 1 cast to a char. It had been used in an older SquidLib version to mark chars in a 2D char array before they were edited, but wasn't used currently, so the checks for `'\1'` weren't helpful in LineTools, FOV, etc.