Match strange old behavior from SquidLib.
IntIntOrderedMap.putIfAbsent() (in SquidLib, not jdkgdxds) had a bug where it would sometimes return the current value and sometimes return the previous value for a key. Now it matches Map and always returns the previous value, which is quickly fixed here. The results in MarkovText are a lot better now.