BREAKING: StringTools.joinAlt() renamed.
Now the version that joins boolean values densely is joinDense/appendJoinedDense, and the version that appends long values so Java can read them is joinReadably/appendJoinedReadably (the last is new). Some overloads that might have clashed in practice are gone, such as joining a Collection of CharSequence when joining an Iterable of Object works fine for that case.