Thank you, IntelliJ IDEA, for this refactoring.
When accessing a Coord's x or y position, these are now encapsulated into x() and y() functions, though I'm not sure if this has to remain so. This allows x and y to become shorts (which they were in practice anyway) and to get a precalculated, more involved hashCode() without needing more memory.