Odd! A small potential bug in DijkstraMap.
When two goals are orthogonally adjacent and a path approaches them such that both goals are adjacent by Chebyshev metric to the nearest cell in the path, the next cell the path chooses is not necessarily the closer goal when the pathfinder is using Euclidean metric. The closer goal does seem to be picked for Manhattan, and there is no better choice for Chebyshev, but for Euclidean it can pick the wrong goal.