OK, found some bugs in anim8-gdx.
PaletteReducer.analyzeHueWise() sometimes produces duplicate colors, meaning it isn't using as many colors as it says it is. This may be related to an issue where a single pixel on one frame of an animation flashes either white or transparent, then goes back to normal on the next frame. It will be hard to track down...