NAAAAAAAAAAAAAAATE!!!
How is Kryo so astonishingly broken??? Writing an int and reading an int works, wow, great job. But, writing an int length, writing that many bytes, and then reading in an int (the length???) gives some completely unrelated number. Since this is a length, it allocates that many bytes. And this number is sometimes huge (half a billion) or sometimes smaller than what we need to read (buffer underflow). Either I'm doing something wrong somewhere, or I need to stop suggesting Kryo as a valid option for ser/deser. It's in version 5.6.2 and is still incredibly unreliable. AAAAGH.