Make that 128x128!
In order to adequately map each grayscale value (with very few exceptions) to a different sized range, we need a 128x128 block size. This doesn't change size at 63, the very center (127 and 128 are the same), or 192, but changes on all other sizes. This code currently won't adapt for block sizes other than 128x128... It should be possible, but hard, to get it to map to other sizes.