Some corrections in Biome!
Some older code, such as equals() and hashCode(), treated the color as a float rather than an int (it is an int now). This probably would have made some comparisons incorrect... The name for empty space is also "Space" now, not "Empty".

This also updates juniper to 0.6.7, jdkgdxds-interop to its latest version 1.8.1.2, and sets the Sonatype host for the next release to the right one for now.