I have no idea why this acts how it does.
The Hyperelliptical projection with alpha=0, kappa=1 always has poles that are two cells wide, and equator-extremes that are one cell tall. I think this is because of some error in my calculations, but I don't know nearly enough of the math involved to do this. The diamond shapes aren't the data I need anyway... They need to be squares, with 0,0 being the south pole and width-1,height-1 the north pole.